+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## Added
+- Some new drinks from Season 2
+
+## Changed
+- Cataclyst charges in Character Frame updated for Season 2 Essence
+
 ## Bugfix
-- When a Quest has only one quest reward, sometimes it was not automatically completed
+- Sometimes a LUA-Error because of quest interaction happend